+++ robot-command/src/main/java/org/obolibrary/robot/CommandLineHelper.java
@@ -470,6 +470,7 @@ public class CommandLineHelper {
       throws IllegalArgumentException, IOException {
     List<String> inputOntologyPaths = getOptionalValues(line, "input");
     List<String> inputOntologyIRIs = getOptionalValues(line, "input-iri");
+    String inputFormat = getOptionalValue(line, "input-format");
 
     int check = inputOntologyPaths.size() + inputOntologyIRIs.size();
     if (check > 1) {
@@ -477,17 +478,9 @@ public class CommandLineHelper {
     }
 
     if (!inputOntologyPaths.isEmpty()) {
-      if (catalogPath != null) {
-        return ioHelper.loadOntology(inputOntologyPaths.get(0), catalogPath);
-      } else {
-        return ioHelper.loadOntology(inputOntologyPaths.get(0));
-      }
+      return ioHelper.loadOntology(inputOntologyPaths.get(0), catalogPath, inputFormat);
     } else if (!inputOntologyIRIs.isEmpty()) {
-      if (catalogPath != null) {
-        return ioHelper.loadOntology(IRI.create(inputOntologyIRIs.get(0)), catalogPath);
-      } else {
-        return ioHelper.loadOntology(IRI.create(inputOntologyIRIs.get(0)));
-      }
+      return ioHelper.loadOntology(IRI.create(inputOntologyIRIs.get(0)), catalogPath, inputFormat);
     } else {
       // Both input options are empty
       throw new IllegalArgumentException(missingInputError);
@@ -913,6 +906,11 @@ public class CommandLineHelper {
     o.addOption("v", "verbose", false, "increased logging");
     o.addOption("vv", "very-verbose", false, "high logging");
     o.addOption("vvv", "very-very-verbose", false, "maximum logging, including stack traces");
+    o.addOption(
+        null,
+        "input-format",
+        true,
+        "format of the input ontology: obo, owl, ttl, owx, omn, ofn, json");
     o.addOption(null, "catalog", true, "use catalog from provided file");
     o.addOption("p", "prefix", true, "add a prefix 'foo: http://bar'");
     o.addOption("P", "prefixes", true, "use prefixes from JSON-LD file");
@@ -1073,21 +1071,22 @@ public class CommandLineHelper {
   public static List<OWLOntology> getInputOntologies(IOHelper ioHelper, CommandLine line)
       throws IllegalArgumentException, IOException {
     List<OWLOntology> inputOntologies = new ArrayList<>();
+    String inputFormat = getOptionalValue(line, "input-format");
     // Check for input files
     List<String> inputOntologyPaths = getOptionalValues(line, "input");
     for (String inputOntologyPath : inputOntologyPaths) {
-      inputOntologies.add(ioHelper.loadOntology(inputOntologyPath));
+      inputOntologies.add(ioHelper.loadOntology(inputOntologyPath, true, inputFormat));
     }
     // Check for input IRIs
     List<String> inputOntologyIRIs = getOptionalValues(line, "input-iri");
     for (String inputOntologyIRI : inputOntologyIRIs) {
-      inputOntologies.add(ioHelper.loadOntology(IRI.create(inputOntologyIRI)));
+      inputOntologies.add(ioHelper.loadOntology(IRI.create(inputOntologyIRI), null, inputFormat));
     }
     // Check for input patterns (wildcard)
     String pattern = getOptionalValue(line, "inputs");
     if (pattern != null) {
       for (File inputOntologyFile : getFilesByPattern(pattern)) {
-        inputOntologies.add(ioHelper.loadOntology(inputOntologyFile));
+        inputOntologies.add(ioHelper.loadOntology(inputOntologyFile, true, inputFormat));
       }
     }
     return inputOntologies;
@@ -1106,22 +1105,24 @@ public class CommandLineHelper {
   public static List<OWLOntology> getInputOntologies(
       IOHelper ioHelper, CommandLine line, String catalogPath) throws IOException {
     List<OWLOntology> inputOntologies = new ArrayList<>();
+    String inputFormat = getOptionalValue(line, "input-format");
     // Check for input files
     List<String> inputOntologyPaths = getOptionalValues(line, "input");
     for (String inputOntologyPath : inputOntologyPaths) {
-      inputOntologies.add(ioHelper.loadOntology(inputOntologyPath, catalogPath));
+      inputOntologies.add(ioHelper.loadOntology(inputOntologyPath, catalogPath, inputFormat));
     }
     // Check for input IRIs
     List<String> inputOntologyIRIs = getOptionalValues(line, "input-iri");
     for (String inputOntologyIRI : inputOntologyIRIs) {
-      inputOntologies.add(ioHelper.loadOntology(IRI.create(inputOntologyIRI), catalogPath));
+      inputOntologies.add(
+          ioHelper.loadOntology(IRI.create(inputOntologyIRI), catalogPath, inputFormat));
     }
     // Check for input patterns (wildcard)
     String pattern = getOptionalValue(line, "inputs");
     if (pattern != null) {
       File catalogFile = new File(catalogPath);
       for (File inputOntologyFile : getFilesByPattern(pattern)) {
-        inputOntologies.add(ioHelper.loadOntology(inputOntologyFile, catalogFile));
+        inputOntologies.add(ioHelper.loadOntology(inputOntologyFile, catalogFile, inputFormat));
       }
     }
     return inputOntologies;