+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `--input-format` global option [#1038]
+
 ### Fixed
+- Inherit OWLDocumentFormat when performing SPARQL update [#1267]
+- Updated obographs to [version 0.3.2](https://github.com/geneontology/obographs/releases/tag/v0.3.2)
 - Partial fix for prefix injections in OBO files [#1268]
 
 ## [1.9.8] - 2025-05-15